Move top-level tests to their proper test modules
